# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.Error:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
host: 43.138.61.49
info:
  contact:
    email: 3264085417@qq.com
    name: big_dust
  description: MucaveAPI
  termsOfService: http://swagger.io/terrms/
  title: Mucave
  version: 1.1.0
paths:
...
  /post:
    delete:
      consumes:
      - application/json
      description: 根据post_id删除指定帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"删帖成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"删帖失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 删帖
      tags:
      - post
    post:
      consumes:
      - application/json
      description: 上传帖子的各项信息，发布帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 类型
        in: formData
        name: type
        required: true
        type: string
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 文本内容
        in: formData
        name: content
        required: true
        type: string
      - description: 文件组(有数量限制)
        in: formData
        name: file
        type: file
      - description: yse/no(说明是否上传了文件)
        in: query
        name: file_have
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"发布成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"新建帖子失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 发布帖子
      tags:
      - post
    put:
      consumes:
      - application/json
      description: 上传所有内容然后修改所有内容,
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 类型
        in: formData
        name: type
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      - description: 标题
        in: formData
        name: title
        required: true
        type: string
      - description: 文本内容
        in: formData
        name: content
        required: true
        type: string
      - description: 新文件组(有数量限制)
        in: formData
        name: file
        type: file
      - description: 文件地址1
        in: formData
        name: file_path1
        required: true
        type: string
      - description: 文件地址2
        in: formData
        name: file_path2
        required: true
        type: string
      - description: 文件地址3
        in: formData
        name: file_path3
        required: true
        type: string
      - description: 文件地址4
        in: formData
        name: file_path4
        required: true
        type: string
      - description: yse/no(说明是否修改了文件)
        in: query
        name: file
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"修改帖子成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"修改帖子失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 修改帖子
      tags:
      - post
  /post/{id}:
    get:
      consumes:
      - application/json
      description: 通过帖子的id获得某个帖子的详细信息
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定帖子的id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到指定的帖子"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"指定的帖子查询失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 查某个帖子
      tags:
      - post
  /post/comment:
    delete:
      consumes:
      - application/json
      description: 根据comment_id删除指定评论
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论id
        in: query
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"删评论成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"删评论失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 删评论
      tags:
      - post
    post:
      consumes:
      - application/json
      description: 评论指定的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      - description: 评论的内容
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"评论成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"评论失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 评论
      tags:
      - post
  /post/comment_replies/{comment_id}:
    get:
      consumes:
      - application/json
      description: 根据评论的id查询指定评论的所有回复
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定评论的id
        in: path
        name: comment_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到评论的回复"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询评论的回复失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 评论的回复
      tags:
      - post
  /post/comment_reply:
    post:
      consumes:
      - application/json
      description: 给指定评论添加回复
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论的id
        in: query
        name: comment_id
        required: true
        type: integer
      - description: 回复对象的id
        in: formData
        name: object
        required: true
        type: integer
      - description: 回复的内容
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"回复成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"回复失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 回复
      tags:
      - post
  /post/comments/{post_id}:
    get:
      consumes:
      - application/json
      description: 根据帖子的id查询指定帖子的所有评论(不包括评论的回复)
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 指定帖子的id
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到帖子的评论."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"帖子的评论查询失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 帖子的评论
      tags:
      - post
  /post/following:
    get:
      consumes:
      - application/json
      description: 查询关注用户的帖子（所有）
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到关注的用户的帖子的数组"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"查询关注用户的帖子失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 关注用户帖子
      tags:
      - post
  /post/latest:
    get:
      consumes:
      - application/json
      description: 查询最新发布的帖子，返回的开始点和数量由query参数决定
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 开始点
        in: query
        name: start
        required: true
        type: integer
      - description: 需要返回的帖子数量
        in: query
        name: length
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到最新的帖子（数组）"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"最新帖子查询失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 最新的帖子
      tags:
      - post
  /post/likes:
    delete:
      consumes:
      - application/json
      description: 通过帖子id取消点赞
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"取消点赞成功"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"取消点赞失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 取消点赞
      tags:
      - post
    post:
      consumes:
      - application/json
      description: 点赞指定的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"点赞成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"点赞失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 点赞
      tags:
      - post
  /post/recommendations/{type}:
    get:
      consumes:
      - application/json
      description: 查询点赞最多的帖子，返回的开始点和数量由query参数决定，类型有type参数决定
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子类型
        in: path
        name: type
        required: true
        type: string
      - description: 开始点
        in: query
        name: start
        required: true
        type: integer
      - description: 需要返回的帖子数量
        in: query
        name: length
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"查询到推荐的帖子（数组）"}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"推荐帖子查询失败"}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 推荐的帖子
      tags:
      - post
  /post/reply:
    delete:
      consumes:
      - application/json
      description: 根据reply_id删除指定帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 回复id
        in: query
        name: reply_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"删回复成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: '{"msg":"删回复失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 删回复
      tags:
      - post
  /post/search:
    get:
      consumes:
      - application/json
      description: 通过关键词搜索标题有关键词的帖子
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 关键词
        in: query
        name: query_string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"搜索成功."}'
          schema:
            $ref: '#/definitions/handler.Response'
        "410":
          description: '{"msg":"搜索失败."}'
          schema:
            $ref: '#/definitions/handler.Error'
      summary: 搜索帖子
      tags:
      - post
  /post/whether_like:
    get:
      consumes:
      - application/json
      description: 通过帖子id查询是否已经点赞
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"msg":"yes"}'
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 是否已赞
      tags:
      - post
...
schemes:
- http
swagger: "2.0"